Upgrade Genocs Blazor WebAssembly Template to version 1.1.0
This commit updates the Genocs Blazor WebAssembly Template to version 1.1.0. The previous version was 1.0.0, and this update brings stability improvements and bug fixes. It is recommended to use the latest available version of the template. Please refer to the [nuget.org](https://www.nuget.org/packages/Genocs.BlazorWasm.Template/) page for the latest version information.